refactored out update code into own script
invoke manually or via cron job with php -f update.php
